Rewrite compatibility test suite in ESM and run it with AVA
