+++ package.json
@@ -64,7 +64,6 @@
     "is-ci": "3.0.1",
     "licensee": "10.0.0",
     "markdownlint-cli": "0.35.0",
-    "mocha": "9.2.2",
     "nve": "16.0.0",
     "prettier": "3.0.0",
     "publint": "0.1.15",
@@ -110,8 +109,8 @@
     "mutation:integration": "stryker run stryker.integration.config.js",
     "mutation:unit": "stryker run stryker.unit.config.js",
     "test": "npm run test:unit",
-    "test:compat": "mocha test/compat/**/*.test.cjs",
-    "test:compat-all": "nve 14.18.0,16.13.0,18.0.0,19.0.0,20.0.0 mocha test/compat/**/*.test.cjs",
+    "test:compat": "ava test/compat/**/*.test.js",
+    "test:compat-all": "nve 14.18.0,16.13.0,18.0.0,19.0.0,20.0.0 npm run test:compat --ignore-scripts",
     "test:e2e": "ava test/e2e/**/*.test.js --timeout 1m",
     "test:integration": "ava test/integration/**/*.test.js --timeout 1m",
     "test:unit": "ava test/unit/**/*.test.js",
